backend/requirements.txt
fastapi
uvicorn
pymupdf
google-generativeai
python-multipart
python-dotenv